+++ internal/localhttp/config.go
@@ -1,4 +1,4 @@
-package uibridge
+package localhttp
 
 import "time"
