+++ internal/localhttp/generate_swagger.go
@@ -1,4 +1,4 @@
-package uibridge
+package localhttp
 
 import (
 	"fmt"
